fix!: move pkg to internal
