+++ internal/config/v1/config.go
@@ -6,9 +6,9 @@ import (
 	"time"
 
 	"github.com/prometheus/client_golang/prometheus"
-	"github.com/raffis/mongodb-query-exporter/v2/collector"
-	"github.com/raffis/mongodb-query-exporter/v2/config"
-	"github.com/raffis/mongodb-query-exporter/v2/x/zap"
+	"github.com/raffis/mongodb-query-exporter/v3/internal/collector"
+	"github.com/raffis/mongodb-query-exporter/v3/internal/config"
+	"github.com/raffis/mongodb-query-exporter/v3/internal/x/zap"
 
 	"go.mongodb.org/mongo-driver/mongo/options"
 )
