+++ internal/config/v2/config.go
@@ -1,4 +1,4 @@
-package v3
+package v2
 
 import (
 	"context"
@@ -8,46 +8,40 @@ import (
 	"time"
 
 	"github.com/prometheus/client_golang/prometheus"
-	"github.com/raffis/mongodb-query-exporter/v2/collector"
-	"github.com/raffis/mongodb-query-exporter/v2/config"
-	"github.com/raffis/mongodb-query-exporter/v2/x/zap"
+	"github.com/raffis/mongodb-query-exporter/v3/internal/collector"
+	"github.com/raffis/mongodb-query-exporter/v3/internal/config"
+	"github.com/raffis/mongodb-query-exporter/v3/internal/x/zap"
 
 	"go.mongodb.org/mongo-driver/mongo/options"
 )
 
-// Configuration v3.0 format
+// Configuration v2.0 format
 type Config struct {
-	Bind         string
-	MetricsPath  string
-	Log          zap.Config
-	Global       Global
-	Servers      []*Server
-	Aggregations []*collector.Aggregation
+	Bind        string
+	MetricsPath string
+	Log         zap.Config
+	Global      Global
+	Servers     []*Server
+	Metrics     []*Metric
 }
 
-// Global config
 type Global struct {
 	QueryTimeout      time.Duration
 	MaxConnections    int32
-	DefaultCache      time.Duration
+	DefaultCache      int64
 	DefaultMode       string
 	DefaultDatabase   string
 	DefaultCollection string
 }
 
-// Aggregation defines what aggregation pipeline is executed on what servers
-type Aggregation struct {
-	Servers    []string
-	Cache      time.Duration
-	Mode       string
-	Database   string
-	Collection string
-	Pipeline   string
-	Metrics    []Metric
-}
-
-// Metric defines how a certain value is exported from a MongoDB aggregation
+// Metric defines an exported metric from a MongoDB aggregation pipeline
 type Metric struct {
+	Servers       []string
+	Cache         int64
+	Mode          string
+	Database      string
+	Collection    string
+	Pipeline      string
 	Name          string
 	Type          string
 	Help          string
@@ -116,7 +110,7 @@ func (conf *Config) Build() (*collector.Collector, error) {
 	c := collector.New(
 		collector.WithConfig(&collector.Config{
 			QueryTimeout:      conf.Global.QueryTimeout,
-			DefaultCache:      conf.Global.DefaultCache,
+			DefaultCache:      time.Duration(conf.Global.DefaultCache) * time.Second,
 			DefaultMode:       conf.Global.DefaultMode,
 			DefaultDatabase:   conf.Global.DefaultDatabase,
 			DefaultCollection: conf.Global.DefaultCollection,
@@ -158,38 +152,32 @@ func (conf *Config) Build() (*collector.Collector, error) {
 		}
 	}
 
-	if len(conf.Aggregations) == 0 {
-		l.Sugar().Warn("no aggregations have been configured")
+	if len(conf.Metrics) == 0 {
+		l.Sugar().Warn("no metrics have been configured")
 	}
 
-	for i, aggregation := range conf.Aggregations {
-		opts := &collector.Aggregation{
-			Servers:    aggregation.Servers,
-			Cache:      aggregation.Cache,
-			Mode:       aggregation.Mode,
-			Database:   aggregation.Database,
-			Collection: aggregation.Collection,
-			Pipeline:   aggregation.Pipeline,
-		}
+	for _, metric := range conf.Metrics {
+		err := c.RegisterAggregation(&collector.Aggregation{
+			Servers:    metric.Servers,
+			Cache:      time.Duration(metric.Cache) * time.Second,
+			Mode:       metric.Mode,
+			Database:   metric.Database,
+			Collection: metric.Collection,
+			Pipeline:   metric.Pipeline,
+			Metrics: []*collector.Metric{
+				{
+					Name:          metric.Name,
+					Type:          metric.Type,
+					Help:          metric.Help,
+					Value:         metric.Value,
+					OverrideEmpty: metric.OverrideEmpty,
+					EmptyValue:    metric.EmptyValue,
+					ConstLabels:   metric.ConstLabels,
+					Labels:        metric.Labels,
+				},
+			},
+		})
 
-		for _, metric := range aggregation.Metrics {
-			opts.Metrics = append(opts.Metrics, &collector.Metric{
-				Name:          metric.Name,
-				Type:          metric.Type,
-				Help:          metric.Help,
-				Value:         metric.Value,
-				OverrideEmpty: metric.OverrideEmpty,
-				EmptyValue:    metric.EmptyValue,
-				ConstLabels:   metric.ConstLabels,
-				Labels:        metric.Labels,
-			})
-		}
-
-		if len(conf.Aggregations) == 0 {
-			l.Sugar().Warn("no metrics have been configured for aggregation_%d", i)
-		}
-
-		err := c.RegisterAggregation(aggregation)
 		if err != nil {
 			return c, err
 		}
